Split up genesis_utils
Also fixed config file location
