+++ crates/genesis_brain/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 license = "MIT"
 
 [dependencies]
-genesis_util = {path = "../genesis_util/"}
+genesis_newtype = {path = "../genesis_newtype/"}
 genesis_maths = {path = "../genesis_maths/"}
 rand = { workspace = true }
 rand_distr = { workspace = true }
