+++ crates/genesis_color/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+authors = ["Ben Askew-Gale <bengale2007@googlemail.com>"]
+name = "genesis_color"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+
+[dependencies]
+genesis_newtype = { path = "../genesis_newtype/"}
+ndarray = {workspace = true}
