+++ crates/genesis_genome/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 license = "MIT"
 
 [dependencies]
-genesis_util = {path = "../genesis_util/"}
+genesis_newtype = {path = "../genesis_newtype/"}
 rand = { workspace = true}
 thiserror = { workspace = true}
 bitvec = { workspace = true}
