+++ crates/genesis_newtype/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 authors = ["Ben Askew-Gale <bengale2007@googlemail.com>"]
-name = "genesis_util"
+name = "genesis_newtype"
 version = "0.1.0"
 edition = "2021"
 license = "MIT"
@@ -8,7 +8,3 @@ license = "MIT"
 [dependencies]
 rand = {workspace = true}
 thiserror = {workspace = true}
-nalgebra = {workspace = true}
-glam = {workspace = true}
-ndarray = {workspace = true}
-num = {workspace = true}
