--- crates/genesis_util/src/util_error.rs
@@ -1,10 +1,0 @@
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum GenesisUtilError {
-    #[error("Weight must be between -1 and 1.")]
-    InvalidWeight,
-
-    #[error("Probability must be between 0 and 1.")]
-    InvalidProbability,
-}
