--- genesis_game/assets/config/genesis.toml
@@ -1,130 +1,0 @@
-start_num = 10
-minimum_number = 5
-initial_synapse_count = 3
-mutations = 3
-start_energy = 800
-lowest_energy_limit = 600
-rotation_cost = [
-    0.02,
-    0.1,
-]
-translation_cost = [
-    0.02,
-    0.1,
-]
-unit_size_cost = 0.02
-world_energy = 30000
-plant_energy_per_unit = 2
-plant_size_range = [
-    10.0,
-    30.0,
-]
-
-[[spawners]]
-centre = [
-    0.0,
-    0.0,
-]
-radius = 500.0
-
-[spawners.dist]
-name = 'normal'
-a = 0.0
-b = 1.0
-
-[attributes]
-hatch_age = [
-    10.0,
-    30.0,
-    15,
-]
-adult_age = [
-    50.0,
-    70.0,
-    20,
-]
-death_age = [
-    600.0,
-    700.0,
-    50,
-]
-mutation_probability = [
-    0.01,
-    0.35,
-    100,
-]
-max_speed = [
-    100.0,
-    500.0,
-    100,
-]
-max_rotation = [
-    10.0,
-    30.0,
-    20,
-]
-eye_range = [
-    200.0,
-    700.0,
-    100,
-]
-eye_angle = [
-    360.0,
-    30.0,
-    100,
-]
-internal_timer_boundary = [
-    -0.5,
-    0.5,
-    20,
-]
-lay_egg_boundary = [
-    0.0,
-    0.8,
-    30,
-]
-want_to_grow_boundary = [
-    -0.5,
-    0.5,
-    20,
-]
-eating_boundary = [
-    -0.5,
-    0.5,
-    20,
-]
-cost_of_thought = [
-    0.001,
-    0.003,
-    10,
-]
-cost_of_eating = [
-    0.3,
-    0.2,
-    10,
-]
-offspring_energy = [
-    0.5,
-    1.0,
-    100,
-]
-mouth_width = [
-    0.3,
-    0.6,
-    20,
-]
-hatch_size = [
-    20.0,
-    35.0,
-    15,
-]
-max_size = [
-    80.0,
-    100.0,
-    20,
-]
-growth_rate = [
-    0.05,
-    0.1,
-    20,
-]
