+++ justfile
@@ -11,4 +11,4 @@ test PACKAGE:
     @cargo nextest run -p {{PACKAGE}}
 
 clean:
-    @rm assets/config/genesis.toml
+    @rm config/genesis.toml
